@aead 2016-12-20
This is a duplicate of  #3420 - as soon as this PR is merged, the PR #3420 should probably be closed